requirements.txt
torch
transformers
datasets
peft
bitsandbytes
accelerate
trl
pyyaml
tqdm
